Move the halls components to it's components directory
